+++ py_src/stuck_two_day/test_cases/not_stuck.json
